+++ .github/workflows/tests.yml
@@ -1,0 +1,22 @@
+name: Krane pipeline
+
+on:
+  pull_request:
+    branches: [ master ]
+
+jobs:
+  run_tests:
+    name: Tests
+    runs-on: ubuntu-latest
+    timeout-minutes: 20
+    steps:
+      - uses: actions/checkout@v2
+
+      - name: Build Krane
+        run: go build -v ./...
+        working-directory: './'
+
+      - name: Run tests
+        run: go test -v ./...
+        working-directory: './'
+
