+++ build_configuration.go
@@ -1,0 +1,6 @@
+package main
+
+type BuildConfiguration struct {
+	Images  []Image `yaml:"build"`
+	Threads int     `yaml:"threads"`
+}
